[autotest] Eliminate a race during provisioning.
During provisioning, we make sure that the ".labmachine" file
exists, and then run the 'autoreboot' upstart job, which depends on
that file existing.  However, most of the time, the '.labmachine'
already exists.  Re-running the 'autoreboot' job is safe, but
Upstart would fail if two 'start autoreboot' commands happened
at the same time.  This race meant that sometimes provisioning
would fail.

This changes the command run during post-provisioning checks to
avoid the race condition.

BUG=chromium:514802
TEST=(will be tested during push to prod)

Change-Id: I4bb57121abd0a3f0f8cb7ed3cf79faea63b26941
Reviewed-on: https://chromium-review.googlesource.com/315206
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>