[autotest] Schedule FAFT on ToT firmware
Use device labels to select if a DUT needs to run FAFT on ToT
or branched firmware.

A DUT which has the original "pool:faft-test" label runs FAFT on
the firmware branch. A DUT which has the new "pool:faft-test-tot"
label runs FAFT on ToT. These labels can be easily changed for
test requirement. For example, when we branch the firmware code
from ToT to a new branch, we can relabel the DUT from
"pool:faft-test-tot" to "pool:faft-test".

BUG=chromium:565668
TEST=site_utils/suite_scheduler/suite_scheduler.py --sanity

Change-Id: Iada09fdf838fb62265bc735e92d39826b6311d28
Reviewed-on: https://chromium-review.googlesource.com/315811
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>