android_xmlrpc_server: Add a loop for ssid discovery
Add a while loop to wait for the SSID to be found in the scan results.
This prevents us from failing discover prematurely. Use the provided
discovery_timeout to wait for the ssid to appear.

While there:
Remove couple of unused constants from AndroidXmlRpcDelegate.

BUG=b:25695720
TEST=`test_that --board=panther android1758-wifi-test-station-2.cros
      network_WiFi_MissingBeacons --args use_wpa_cli=false`

      Change-Id: I806a78f59f8c8185be19763a52c4535754c7fbc0

Change-Id: I2d2d00f84356381ed0e62487a2909b390d63b0f7
Reviewed-on: https://chromium-review.googlesource.com/315774
Commit-Ready: Roshan Pius <rpius@chromium.org>
Tested-by: Roshan Pius <rpius@chromium.org>
Reviewed-by: Ang Li <angli@google.com>
Reviewed-by: Roshan Pius <rpius@chromium.org>