autotest (wifi): Add loop to check for cleared IP on disconnect
Add a loop to check for a cleared IP address in the DisconnectClearsIP
test.  This change is needed due to a variable delay added in the
DropConnection function.  The delay is a result of switching from an
async to sync call when killing dhcpcd.

The number of loops for the test was set after running several
iterations of the test.

BUG=chrome-os-partner:48228
TEST=test_that -b daisy chromeos1-dev-host5.cros network_WiFi_DisconnectClearsIP

Change-Id: I6e078603051c31dabc910e3c9bb550ae26652d41
Reviewed-on: https://chromium-review.googlesource.com/315737
Commit-Ready: Rebecca Silberstein <silberst@chromium.org>
Tested-by: Rebecca Silberstein <silberst@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>