[autotest] Fix provisioning via test_that
Provisioning currently assumes you are running in the lab
environment with an AFE. It used to use the AFE to verify the
right image is installed on the device. This logic is
unnecessary as machine_install would have failed if it could
not properly update the lab AFE anyways.

BUG=chromium:463228
TEST=test_that chromeos4-row10-rack11-host17.cros dummy_PassServer \
-i peppy-release/R49-7712.0.0

Change-Id: If78166a37936c097981dbc110258947270f5ebf1
Reviewed-on: https://chromium-review.googlesource.com/316476
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>