kernel_FirmwareRequest: new test for kernel's fw request features
This test will run the firmware request kernel self test (from
upstream). This tests that the request_firmware() and
request_firmware_nowait() kernel APIs are somewhat sane. It tries to
load the empty filename ("") as well as a small toy firmware, and checks
that it matches. It also makes sure a non-existent firmware cannot be
found.

We rerun the same test several times to increase the probability of
catching errors.

BUG=chromium:557355
TEST=test_that kernel_FirmwareRequest

Change-Id: Ibbe0d89569e2c4eaba6c39f2b2e78b107157e7e7
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316507
Reviewed-by: Kees Cook <keescook@chromium.org>