Implement policy_URLWhitelist AutoTest
This test verifies the behavior of Chrome OS for a range of valid
values of the URLWhitelist user policy, using three unique
test cases: NotSet, SingleWhitelistedFile, MultipleWhitelistedFiles.

When the policy is undefined (case=NotSet), All of the URLs are blocked.
When set to one or more URLs, navigation to the URLs
with the same domain as any of the listed URL/domains is allowed.

BUG=chromium:554434
TEST=Run test_that <IP> policy_URLWhitelist

Change-Id: I0c2147834fe7a61eac516a17851ad7bd0cb860a9
Reviewed-on: https://chromium-review.googlesource.com/317563
Commit-Ready: Krishna Gavini <krishnargv@chromium.org>
Tested-by: Krishna Gavini <krishnargv@chromium.org>
Reviewed-by: Scott Cunningham <scunningham@chromium.org>