hardware_PerfCounterVerification: Loosen failure bounds; add perf tracking
Some non-arm boards are observed to have weaker correlation in
cycles ~ loops, so reduce the test failure threshold to match that
previously used for ARM.

Also output the values to the performance dashboard so that we can better
monitor changes in the values.

Finally, add a BUG_TEMPLATE to the control file so that we get notified
of failures.

BUG=none
TEST=Nope. Doesn't change anything material.

Change-Id: Icd988bfe44d6fbbb487d8a4d7c6abfb5a6e0cdbd
Reviewed-on: https://chromium-review.googlesource.com/317670
Commit-Ready: David Sharp <dhsharp@chromium.org>
Tested-by: David Sharp <dhsharp@chromium.org>
Reviewed-by: Emma Rapati <rapati@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>