[autotest] Fix a bug in machine list comparison.
CL 315230 changed the "machines" variable in a job run to a list of
dictionaries, e.g.,
[{'host_attributes': {}, 'hostname': '100.96.51.226'}]
This breaks the logic in _make_parallel_wrapper, which it checks if
the autoserv run needs to fork based on the comparison of machines and
self.machines, which is still a list of hostnames.

This change converted the list of dictionaries back to a list of hostnames.

BUG=chromium:568802
TEST=local run a client side test, make sure status.log is not empty

Change-Id: Id7be1852b13418bf1b23caaa202418f3381fe51a
Reviewed-on: https://chromium-review.googlesource.com/317579
Commit-Ready: Dan Shi <dshi@chromium.org>
Tested-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Fang Deng <fdeng@chromium.org>