autotest: firmware_ECBootTime race fix
Some systems run the EC code out of ram, rather than flash. There is a
new requirement that these system have their power cut to ensure that
the ram version of the read-only code is not tampered with. So these
systems do a double boot of the EC, once by themselves and once by
asking the PMIC to power cycle the EC.

The ECBootTime test asks the EC to do a "reboot ap-off" to reboot. It
then looks for the string "Inits done" on the EC console. It then sends
the command "powerbtn" to get the EC to boot the AP.

On systems which do this double boot the "Inits done" string is shown by
the first boot. The "powerbtn" may or may not be sent whilst the EC is
being power cycled by the PMIC, and hence it might be lost, leaving the
AP powered down. The rest of the test then times out and the test fails.

Using a string which is shown only in the second boot fixes the race
between the "powerbtn" command and the second boot.

It has the nice side effect of also making the measurement on "EC boot
time", which appears to want to measure the time between the "powerbtn"
command and the first spark of life of the AP, more accurate.

BRANCH=None
BUG=chrome-os-partner:48564
TEST=successfully run "test_that -b kunimitsu $IP firmware_ECBootTime"

Change-Id: I471074fd4f0168bb77191cf74e51f49db095fc8d
Signed-off-by: Icarus Sparry <icarus.w.sparry@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/318630
Commit-Ready: Icarus W Sparry <icarus.w.sparry@intel.com>
Tested-by: Icarus W Sparry <icarus.w.sparry@intel.com>
Tested-by: Naveenkrishna Ch <naveenkrishna.ch@intel.com>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>