When doing an ADB reboot make sure device is ready.
Check that device state before returning that the device
reboot is complete.  A test should have to pool for the device
waiting for the correct state.

I was also unclear why adb_host does not take advantage of
adb 'wait-for-device' as this is only used in initialize, but
tests will want to reboot a device in the middle of a test.

TEST=Manual; Ran it
BUG=chromium:570437

Change-Id: I6f6bc6ff137d7b501c20e75d88b265695bf37ce4
Reviewed-on: https://chromium-review.googlesource.com/318653
Commit-Ready: Kris Rambish <krisr@chromium.org>
Tested-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>