[autotest] Pass in connectivity_class to create_host
This change allows a caller of create_host to predetermine the
connectivity class that should be used when talking to this host.

This is useful when running a VM on localhost as this should be
using an SSHHost and not running directly on the system.

BUG=b:26231815
TEST=local run.

Change-Id: I58d2df8abdc7fcf5657b31c6fb990d7490610de3
Reviewed-on: https://chromium-review.googlesource.com/318754
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>