brillo: Rename brillo-pts to brillo-bvt
PTS is supposed to represent all the tests required for a partner
to claim a board is brillo certified. Following converstations
with partner eng, we have decided to rename the suite to brillo-bvt
as PTS is a superset of these tests and the autoupdate E2E test.

BUG=None
TEST=None

Change-Id: I309f3f3b17c8d98500937b769c7100fa48b9c8f5
Reviewed-on: https://chromium-review.googlesource.com/321111
Commit-Ready: Simran Basi <sbasi@chromium.org>
Tested-by: Simran Basi <sbasi@chromium.org>
Reviewed-by: Mohammed Habibulla <moch@google.com>
Reviewed-by: Dan Shi <dshi@chromium.org>