Rework servo power state test to find more bugs.
Recently, it was discovered that veyron boards were passing the
platform_ServoPowerStateController, even though in practice the
board couldn't power on after powering off twice in a row.  This
change reorders the test to make sure that the test for "power off
twice in a row works" is followed by a test for "power on works".

BUG=chromium:576379
TEST=run against link and veyron DUTs in the lab

Change-Id: I1b781fb094d0590ecda0a1c3560653f00a90fb73
Reviewed-on: https://chromium-review.googlesource.com/321432
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>