autotest (wifi): make DisconnectReason work on whirlwind
The switch_ap subcase of DisconnectReason sets up APs on
two different 5 GHz channels, and switches between them.
On whirlwind, we prefer not to use multiple channels
within the same band, as doing so requires use of the
AUX radio.

There's no particular need for both APs to be in the
same band for this test. So, change ALT_CHANNEL to
be in the 2.4 GHz band.

BUG=chromium:576950
TEST=network_WiFi_DisconnectReason.switch_ap

Change-Id: Ifa6552f6e1d90b6ab552db643801291e6ebc015c
Reviewed-on: https://chromium-review.googlesource.com/321388
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Rebecca Silberstein <silberst@chromium.org>