autotest (wifi): make MultiAuth work on whirlwind
The MultiAuth test sets up SSIDs on two different 2.4 GHz
channels. On whirlwind, we prefer not to use multiple
channels within the same band, as doing so requires use
of the AUX radio.

There's no particular need for the two SSIDs to be
on different channels for this test. So, run both
SSIDs on the same channel.

Alternatively, we could spread the SSIDs across bands.
However, that would make the test fail on single-band
clients.

BUG=chromium:576996
TEST=network_WiFi_MultiAuth (veyron_speedy+whirlwind)

Change-Id: If34e9606bf7cd3c7239dd531da7f734c2a0a0ad7
Reviewed-on: https://chromium-review.googlesource.com/321861
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>