Increase time allowed for servod to start in deployment_test.
Sometimes, the "deployment_test" script fails with "Connection
refused" from the servo.  Looking at the logs, there doesn't
seem to be anything wrong with the servo as such, but it's possible
that we're not allowing servod enough time to get started.

This increases the time allowance from 4 seconds to 10 seconds.

BUG=chromium:577324
TEST=None

Change-Id: Ibb7f3e8a03a64760758ceb29b0ac742ad0c9f3db
Reviewed-on: https://chromium-review.googlesource.com/321950
Commit-Ready: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>