faft: Added USB PD VBUS change test
Created new FAFT USB PD test that causes VBUS change requests
and verifies that the voltage change is successful. This test
relies on both the DUT and Plankton to execute.

BRANCH=none
BUG=chrome-os-partner:49305
TEST=Manual
Tested using Samus and Plankton with workstation as localhost. Verfied
the VBUS changes work as expected. In addition, modified the tolerance
ranges to force failures and verify that the error paths worked
as expected.

Change-Id: I9041efa059210705890bab204ae48b40f6072991
Signed-off-by: Scott Collyer <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/321862
Reviewed-by: Todd Broch <tbroch@chromium.org>