graphics_dEQP: add debug option
It is sometimes useful to save additional information about what
happened during a test. Some of this data is currently collected
by autotest, but it is difficult to associate it with a specific
test and to correlate data from multiple sources.

This patch adds a debug option to collect dmesg, stdout and stderr
data for each test. This data is written in JSON format to
<test-name>.debug in the same directory as the individual dEQP
test log files on the DUT.

BUG=None
TEST=test_that --args="debug=True filter=dEQP-GLES3.info" <ip> graphics_dEQP.filter_args
     Look for <test-name>.debug files in /tmp/dEQP-GLES3.info-logs
     on the DUT.

Change-Id: Icd6d78ac369d288f89f9fb69dcaa60b8e3278c35
Signed-off-by: Mike Mason <michael.w.mason@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/322225
Commit-Ready: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>