autotest: Repair the firmware earlier
Firmware is easy to be corrupted when running FAFT. The current
execution sequence of the repair functions makes the repair job
take 1 hour to recover the corrupted firmware. We should run
the firmware repair function earlier. As the function is only
executed on the dedicated pool:faft-test, this change won't
affect other pools.

BUG=chromium:577924
TEST=None, trivial change.

Change-Id: I750bc5190c398a3d95a5ceee7eaac095806057e8
Reviewed-on: https://chromium-review.googlesource.com/322247
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>