Experiment running WebRTC Camera tests on dedicated machines.
We have noticed that WebRTC tests occasionaly fail. We suspect that
the failure is due to lack of enough lighting on some devices in the
lab.

In order to make these tests reliable, we could run them on dedicated
machines that are places where there is sufficient light.

This CL creates a new suite for these tests to run. As we get more
reliable results this way, we will expand the list of devices.

BUG=None.
TEST=Wait for lab.

Change-Id: I0c52d6b04cf1cc28cb09caa74b3f1d242a89ed0f
Reviewed-on: https://chromium-review.googlesource.com/322287
Commit-Ready: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>