MediaRecorder: CrOS test : fix the MediaRecorder's init parameter error.
This CL fixes https://crbug.com/569011 by updating the initialization
parameters for MediaRecorder in the javascript test. In the process
of doing so, the existing test code is updated to correspond to the
current content browser test code. No new test functions where
introduced in this CL.

BUG=chromium:569011

TEST=On Jerry laptop and trybots
Change-Id: I31a10ad262749d16de89cb757c86b89127078692
Reviewed-on: https://chromium-review.googlesource.com/322765
Commit-Ready: Christian Paulin <cpaulin@chromium.org>
Tested-by: Christian Paulin <cpaulin@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>