Increase payload application timeout to 1h.
The update_engine_client command ran from update_image() had a timeout
of 20 minutes. This command blocks until the payload is downloaded and
applied in the DUT. Due to network load or server side load, the
payload application started to take more than 20 minutes, with a
steady but slow download speed.

This CL bumps the limit to 1h and lets the whole test suite timeout
instead of failing to apply the payload.

BUG=chromium:580261
TEST=None

Change-Id: I07bad14ae1a6fd28d22ed371d920953268b8ef05
Reviewed-on: https://chromium-review.googlesource.com/322866
Trybot-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Dan Shi <dshi@google.com>