[autotest] Fix brillo_KernelVersionTest
Fix the brillo_KernelVersionTest to stop checking minor version
numbers if the actual major version number is higher than the
minimum.  Also fix the test to convert the parsed major and minor
version numbers to ints before doing the comparison, so that
a kernel version '3.9' will fail when the  minimum is '3.10'.

Adds the brillo_KernelVersionTest to the brillo-smoke suite.

BUG=b/26724977
TEST=`test_droid.py emulator-5554 suite:brillo-smoke` runs the
     test and it passes.

Change-Id: I3b1f461771b6ed24dd7ea6ff3f7883097dc665e5
Reviewed-on: https://chromium-review.googlesource.com/323184
Commit-Ready: Steve Fung <stevefung@chromium.org>
Tested-by: Steve Fung <stevefung@chromium.org>
Reviewed-by: Kees Cook <keescook@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>