[autotest] Allow nightly event to specify the hour to run in suite scheduler
This change add a new config `hour` in nightly task. The value specifies the
hour of a day the nightly task should run. This allows the suite jobs for
nightly runs to be distributed across the day to avoid nightly spike on network
load, as each suite job starts with staging images on devserver.

BUG=chromium:580254
TEST=unittest, local run
/usr/local/autotest/site_utils/suite_scheduler/suite_scheduler.py \
    -d /usr/local/autotest/logs -f /usr/local/autotest/test_suite_scheduler.ini \
    -r /tmp/_autotmp_d2jtlh_suite_scheduler -e nightly \
    -i veyron_jerry-release/R49-7834.3.0

DEPLOY=suite scheduler

Change-Id: I8fd8e078b8c5524dc0c0e6453458b730adcdddd3
Reviewed-on: https://chromium-review.googlesource.com/323331
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Dan Shi <dshi@google.com>