faft: Remove firmware_UpdateKernelSubkeyVersion from nightly run
This test is to test the shellball logic. If the active firmware
is not the same as the shellball, this test will fail. So the
nightly FAFT run always fails because it uses a latest firmware,
which is not the same as the shellball.

BUG=chrome-os-partner:44139
TEST=None, trivial change.

Change-Id: I52cb31352aa6b968f3e12701e3046ad11955fc6a
Reviewed-on: https://chromium-review.googlesource.com/323333
Commit-Ready: Wai-Hong Tam <waihong@chromium.org>
Tested-by: Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Shelley Chen <shchen@chromium.org>