[autotest] brillo_StorageWriteSpeedTest
The output from `dd` gets written to stderr.  Since the test needs
to parse the output, redirect it to stdout so that it gets returned
from host.run_output().

Also the emulator's write speed tends to be slower than an actual
device, so add a second control file for running the test in the
smoke suite with a lower min_speed.

BUG=b:26742334
TEST=`./test_droid.py <device> brillo_StorageWriteSpeedTest` passes
TEST=`./test_droid.py --args min_speed=$((1024*1024*1024))
      emulator-5554 suite:brillo-smoke` fails.
TEST=`test_droid.py emulator-5554 suite:brillo-smoke` passes.

Change-Id: I0ea48c02aaa988e3e2a9cdbf0d382bd8d4c96a79
Reviewed-on: https://chromium-review.googlesource.com/323379
Commit-Ready: Steve Fung <stevefung@chromium.org>
Tested-by: Steve Fung <stevefung@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>