autotest: Read iw event log directly from host.
Currently, the IwEventLogger is only able to read the event log file after
it has been copied to the autotest server after the event logger has been
stopped. With this change, we will read the file directly from the host,
to allow multiple checks within a single context, then copy it to the
autotest server after we exit the context.

BUG=None
TEST=Run jetstream sanity autotests which use the iw event logger.

Change-Id: If2a814c90a941e6f46ab031909ce607d01a2c729
Reviewed-on: https://chromium-review.googlesource.com/323771
Commit-Ready: Dane Pollock <danepollock@google.com>
Tested-by: Dane Pollock <danepollock@google.com>
Reviewed-by: Filipe Brandenburger <filbranden@chromium.org>