[autotest] Allow weekly suite to be scheduled at any day of a week.
This change is similar to CL 323331. It enables suite scheduler to run weekly
task at a given day of the week. This allows the weekly suites to be distributed
across the week, rather than all run at once, which causes a spike of load on
devservers and lab network.

BUG=chromium:245937
TEST=unittest, suite_scheduler.py --sanity

Change-Id: I20cdc8abae30a552aed6d3f00a2b2e283a287a52
Reviewed-on: https://chromium-review.googlesource.com/323805
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Fang Deng <fdeng@chromium.org>