Fix policy_* AutoTest regression tests
Most policy_* autotests began failing just before holiday break,
around Dec 18, 2015, in build <R49-7748.0.0, due to a timeout
during OOBE sign-in. The issue was caused to moving the check for
whether cryptohome was mounted immediately after sign-in. The problem
was not seen during local testing because it was a race condition.
The fix moves the cryptohome check after a web page has successfully
loaded.

BUG=chromium:581359
TEST=Run test_that <IP>e:policy_.*

Change-Id: I6dba0e60b6eb1ad9976dd3356e2e091c82404f6f
Reviewed-on: https://chromium-review.googlesource.com/323861
Commit-Ready: Scott Cunningham <scunningham@chromium.org>
Tested-by: Scott Cunningham <scunningham@chromium.org>
Reviewed-by: Jacob Dufault <jdufault@chromium.org>