[autotest] Fix a bug for nightly run in suite scheduler
suite scheduler parses the config file in every loop (about 5min), and merge
the new config to the existing ones. During the merge, deadline is updated.
CL 323331 changed the logic of nightly run for it to be able to trigger at
every hour. This also change how deadline should be updated.

The deadline should be unchanged in merge, otherwise the deadline will be
changed to the next hour, thus no nightly run will be scheduled as deadline
will always be pushed an hour in the future.

This change removes the Merge override in Nightly class so the dealine can
stay unchanged during Merge.

BUG=chromium:581113
TEST=local run, unittest

Change-Id: If602c477692e6a862db4e3ed432fc890ec410bdd
Reviewed-on: https://chromium-review.googlesource.com/324100
Commit-Ready: Dan Shi <dshi@google.com>
Tested-by: Dan Shi <dshi@google.com>
Reviewed-by: Kevin Cheng <kevcheng@chromium.org>