Loop to get a set of good frames.
We have crbug/580736 open tracking the bug that we have get repeated
frames when we try to capture frames.

This CL introduces a loop, when we get repeated frames, we will simply
try again. When that bug is fixed, we should revisit whether we are keeping
the loop or not.

BUG=chromium:580736
TEST=Local.

Change-Id: Iec9dedd700c4c84a7fa85c4aecc6a2a1c15bd968
Reviewed-on: https://chromium-review.googlesource.com/324081
Commit-Ready: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>