autotest: Refactor ping_runner for multi-platform.
There is a desire to add MacOS hosts to WiFi testbeds, so this is an attempt
to extend ping_runner functionality to support multiple platforms which require
different ping arguments and parsing logic, while still keeping the same
external interface. Repurposed PingConfig to only implement very basic input
validation, and command timeout calculation (common between platforms). Added
'PingDelegate' classes which know how to map a PingConfig to arguments, and
implement the (fairly similar) parsing logic of the output. The platform type
is selected when instantiating the ping_runner, and the external interface is
unchanged.

BUG=None
TEST=Ran jetstream_Connect to test linux/chromeos ping_runner behavior is
    unaffected, stub test to verify MacOS ping_runner in a variety of
    scenarios: simple ping, positive/negative test of pings on local subnet,
    positive/negative test of pings to a different subnet.

Change-Id: I5be0a7517618572d03739cf966b17d5170d97025
Reviewed-on: https://chromium-review.googlesource.com/326501
Commit-Ready: Dane Pollock <danepollock@google.com>
Tested-by: Dane Pollock <danepollock@google.com>
Reviewed-by: Christopher Wiley <wiley@chromium.org>