faft: Added USB PD soft reset test
Created a test which executes USB PD soft resets initiated by
both of the port partners. If the connection supports a power
role swap, the power role is swapped, and the test is repeated.

BRANCH=none
BUG=chrome-os-partner:50178
TEST=Manual
Tested with Samus to Plankton. Verfied that soft resets are initiated
and the state transition history can be used to verify expected operation.

Change-Id: Ia0119bd3f9e4dce4cddaf8f91e232a8571439c80
Signed-off-by: Scott <scollyer@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/326785
Commit-Ready: Scott Collyer <scollyer@chromium.org>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>