[Cros WebRtc] Publish specific errors in the test error message.
Although the specific cause of test failure was known, we were
logging that in a file but the test error message that is shown
in wmatrix remained generic.

To get an immediate impression of what caused the test failure,
I'm reporting the error message as the error message.

BUG=None.
TEST=Passes locally.

Change-Id: Ic7e49229b3f563ccbeddd980071c4c7a6f1dfb35
Reviewed-on: https://chromium-review.googlesource.com/327321
Reviewed-by: Rohit Makasana <rohitbm@chromium.org>
Commit-Queue: Mussa Kiroga <mussa@chromium.org>
Trybot-Ready: Mussa Kiroga <mussa@chromium.org>
Tested-by: Mussa Kiroga <mussa@chromium.org>
Reviewed-by: Christian Paulin <cpaulin@chromium.org>