site_gtest_runner: Consider tests that didn't finish as failed.
When the gtest process aborts or crashes during a test, the test output
will contain the beginning of the test run but not the result so it will
remain as 'started' in the gtest_parser. This patch includes the
'started' tests in the the FailedTests() list so tests that abort are
reported in the error logs.

BUG=b:27152312
TEST=Nope.

Change-Id: I6f4887606e453b2fbc1c53d657e2c9b38238f9fe
Reviewed-on: https://chromium-review.googlesource.com/329768
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Paul Hobbs <phobbs@google.com>
Reviewed-by: Simran Basi <sbasi@chromium.org>