kernel_CheckArmErrata: Add new test to check for ARM errata
We recently enabled an ARM Erratum for Cortex-A12.  We want to make sure
that this errata doesn't get lost, even if we later decide to rebase to
a new kernel or we rejitter our configs.

Let's add a test to make sure that the erratum fix is present.  Note that
to test this we need some coprocessor registers that aren't usually
present from userspace, so this depends on a kernel change.  If we
detect that we're on an errata-ridden processor but don't have the
needed kernel infrastructure to check for errata, the test will fail.

This test is designed to support checking for the presence of more
errata on future ARM processors assuming that they can be checked for
using similar methods.

CQ-DEPEND=CL:Ib4685e505ae0411d2dd3e94a0785b5f470523f29
BUG=chrome-os-partner:50137
TEST=test_that kernel_CheckArmErrata and kernel_CheckArmErrata.doctest

Change-Id: I9fbb6f73a8c62886d59753be6223983b1b2c650f
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/329950
Reviewed-by: Brian Norris <briannorris@chromium.org>