deqp: Use RGBA8888 / D24S8 format by default
The surfaceless deqp backend which we use in Chrome OS doesn't specify
a pixel format and runs using an fbo which means that it lets deqp
pick the format.

The format picked for depth is the first in the list of depth formats,
which is GL_DEPTH32F_STENCIL8. Sadly deqp doesn't check that this
format is present (in particular, it is not a requirement for GLES2
implementations, and is only required for GLES3). This means that the
driver rejects the format (rightfully so) and in turn deqp does not
run on these GLES2 stacks.

This patch changes the test to pass -deqp-gl-config-name=rgba8888d24s8
to deqp which forces the use of a D42S8 format, which all GLES
implementations support.

Note that the backends for all the other platforms (android, linux...)
use this format by default already, so by changing this we are making
our deqp testing more comparable with other platforms.

BUG=chromium:532691
TEST=run on alex

Change-Id: I26e432ec7351f636cfe16c265d13de02cd800979
Reviewed-on: https://chromium-review.googlesource.com/330464
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Ilja Friedel <ihf@chromium.org>