AFDO: reduce time for profile collection
Profile collection is taking longer since when this process was started.
This may cause timeout problems in the PFQ builder especially when
benchmarks are retried. This may also be the cause for a weird
linux "perf" failure where it runs out-of-memory at the end of the
profile collection. To ameliorate these problems, I am reducing
the number of iterations for the benchmark that takes the longest time.

BUG=chromium:593086
BUG=chromium:581488
TEST=executed by hand. verified time is reduced.

Change-Id: I70824d0ca47da6d77bd2018e4208681bb2a25fa7
Reviewed-on: https://chromium-review.googlesource.com/331833
Commit-Ready: Luis Lozano <llozano@chromium.org>
Tested-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Caroline Tice <cmtice@chromium.org>