[Autotest] touch_MouseScroll to use actual scrolling again.
Instead of counting from the mouse wheel events, physically scroll
the page (as is done in touch_TouchscreenScroll).

This is because on link and samus, the one value is 2x the other.
It doesn't matter for touch_ScrollDirection, which checkes only
the direction of the scroll.  It /does/ matter for this test,
which checks the absolute value of the delta for a single mouse
click.

TEST=ran on link, samus, and others
BUG=None

Change-Id: Iae2b87f1a2dc20891bd298128097f1d4d7f27bf6
Reviewed-on: https://chromium-review.googlesource.com/331774
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: danny chan <dchan@chromium.org>