[Autotest] Hardcode exception for failing case on touch_UpdateErrors
When link is updated from its very first FSI build, the
touchscreen fails to work on the next boot.  This issue only
comes up in the lab when the AU tests run on a machine before
this test.  The logs from pre-au are not wiped, so this
test fails.

Since there's nothing to be done for link (and since later FSIs
fixed this problem therefore no users see it) hardcode an
exception into this test for link:

- If there are no logs from older builds, proceed as usual.
- When there are logs from an older build, skip those logs
  and look only at the build currently on the machine
- In this subset of the logs, skip the logs from the
  oldest boot.
- If, after these three checks, there are no logs left,
  raise a TestError.

BUG=chromium:593527
TEST=ran on a link in this state.  An introduced failure was
     still found, but the known error was ignored.

Change-Id: I1ba834593d7d9a88461415067e9437d35ddd4b76
Reviewed-on: https://chromium-review.googlesource.com/332283
Commit-Ready: Katherine Threlkeld <kathrelkeld@chromium.org>
Tested-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Kalin Stoyanov <kalin@chromium.org>