Add brillo_Keymaster test.
This test is intended to be a Brillo PTS sanity test for the Brillo
Keymaster hardware requirements. The test will check for a minimal set
of key attributes that are tagged as hardware-backed for all required
key types.

Note: The Brillo Keymaster requirements are based on Android Keymaster
v1.0 requirements but allow exceptions for large key sizes.

BUG=b:27296072
TEST=build, run on brilloemulator

Change-Id: Ied989b7dd7e9bd3fccaffc23b5da499ece6a272e
Reviewed-on: https://chromium-review.googlesource.com/333053
Commit-Ready: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>